_putchar added to function_pointers.h

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,4 @@
-#include <stddef.h>
+#include <stdlib.h>
 #include "function_pointers.h"
 
 /**
@@ -17,7 +17,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned long int i;
 
 	for (i = 0; i < size; i++)
 	{
